ProductCategories in Admin panel Ui Completed now time to go for productManagement in admin

+++ MyNiaSmart/ShopManagement.Application.Contract/ProductStatus/IProductStatusApplication.cs
@@ -7,7 +7,7 @@ namespace ShopManagement.Application.Contract.ProductStatus
     {
         OperationResult Create(CreateProductStatus command);
         OperationResult Edit(EditProductStatus command);
-        EditProductStatus GetDetails(int id);
+        EditProductStatus GetDetails(long id);
         List<ProductStatusViewModel> GetList();
     }
 }
